change c interface to integer8

+++ c/utils.c
@@ -1,9 +1,9 @@
 #include "utils.h"
 #include "math.h"
 
-void dzero(int n, double *a)
+void dzero(int64_t n, double *a)
 {
-  for(int i=0; i<n;i=i+1)
+  for(int64_t i=0; i<n;i=i+1)
   {
     a[i] = 0;
   }
@@ -11,9 +11,9 @@ void dzero(int n, double *a)
 }
 
 
-void czero(int n, CPX *a)
+void czero(int64_t n, CPX *a)
 {
-  for(int i=0; i<n;i=i+1)
+  for(int64_t i=0; i<n;i=i+1)
   {
     a[i] = 0;
   }
@@ -21,7 +21,7 @@ void czero(int n, CPX *a)
 }
 
 
-void comp_err_lap(int n, int pg, int pgt, double *pot, double *potex, double *pottarg,
+void comp_err_lap(int64_t n, int64_t pg, int64_t pgt, double *pot, double *potex, double *pottarg,
       double *pottargex, double *grad, double *gradex, double *gradtarg, double *gradtargex, 
       double *hess, double *hessex, double *hesstarg, double *hesstargex, 
       double *err)
@@ -30,7 +30,7 @@ void comp_err_lap(int n, int pg, int pgt, double *pot, double *potex, double *po
   double r = 0;
   if( pg >= 1)
   {
-    for(int i=0; i<n; i++)
+    for(int64_t i=0; i<n; i++)
     {
       r = r + pow(fabs(potex[i]),2);
       e = e +  pow(fabs(potex[i]-pot[i]),2);
@@ -39,7 +39,7 @@ void comp_err_lap(int n, int pg, int pgt, double *pot, double *potex, double *po
 
   if( pg >= 2)
   {
-    for(int i=0; i<3*n; i++)
+    for(int64_t i=0; i<3*n; i++)
     {
       r = r + pow(fabs(gradex[i]),2);
       e = e +  pow(fabs(gradex[i]-grad[i]),2);
@@ -48,7 +48,7 @@ void comp_err_lap(int n, int pg, int pgt, double *pot, double *potex, double *po
 
   if ( pg >= 3)
   {
-    for(int i=0; i<6*n; i++)
+    for(int64_t i=0; i<6*n; i++)
     {
       r = r + pow(fabs(hessex[i]),2);
       e = e +  pow(fabs(hessex[i]-hess[i]),2);
@@ -57,7 +57,7 @@ void comp_err_lap(int n, int pg, int pgt, double *pot, double *potex, double *po
 
   if( pgt >= 1)
   {
-    for(int i=0; i<n; i++)
+    for(int64_t i=0; i<n; i++)
     {
       r = r + pow(fabs(pottargex[i]),2);
       e = e +  pow(fabs(pottargex[i]-pottarg[i]),2);
@@ -66,7 +66,7 @@ void comp_err_lap(int n, int pg, int pgt, double *pot, double *potex, double *po
 
   if( pgt >= 2)
   {
-    for(int i=0; i<3*n; i++)
+    for(int64_t i=0; i<3*n; i++)
     {
       r = r + pow(fabs(gradtargex[i]),2);
       e = e +  pow(fabs(gradtargex[i]-gradtarg[i]),2);
@@ -75,7 +75,7 @@ void comp_err_lap(int n, int pg, int pgt, double *pot, double *potex, double *po
 
   if ( pgt >= 3)
   {
-    for(int i=0; i<6*n; i++)
+    for(int64_t i=0; i<6*n; i++)
     {
       r = r + pow(fabs(hesstargex[i]),2);
       e = e +  pow(fabs(hesstargex[i]-hesstarg[i]),2);
@@ -86,14 +86,14 @@ void comp_err_lap(int n, int pg, int pgt, double *pot, double *potex, double *po
   return;
 }
 
-void comp_err_helm(int n, int pg, int pgt, CPX *pot, CPX *potex, CPX *pottarg,
+void comp_err_helm(int64_t n, int64_t pg, int64_t pgt, CPX *pot, CPX *potex, CPX *pottarg,
       CPX *pottargex, CPX *grad, CPX *gradex, CPX *gradtarg, CPX *gradtargex, double *err)
 {
   double e = 0;
   double r = 0;
   if( pg >= 1)
   {
-    for(int i=0; i<n; i++)
+    for(int64_t i=0; i<n; i++)
     {
       r = r + pow(cabs(potex[i]),2);
       e = e +  pow(cabs(potex[i]-pot[i]),2);
@@ -102,7 +102,7 @@ void comp_err_helm(int n, int pg, int pgt, CPX *pot, CPX *potex, CPX *pottarg,
 
   if( pg == 2)
   {
-    for(int i=0; i<3*n; i++)
+    for(int64_t i=0; i<3*n; i++)
     {
       r = r + pow(cabs(gradex[i]),2);
       e = e +  pow(cabs(gradex[i]-grad[i]),2);
@@ -111,7 +111,7 @@ void comp_err_helm(int n, int pg, int pgt, CPX *pot, CPX *potex, CPX *pottarg,
 
   if( pgt >= 1)
   {
-    for(int i=0; i<n; i++)
+    for(int64_t i=0; i<n; i++)
     {
       r = r + pow(cabs(pottargex[i]),2);
       e = e +  pow(cabs(pottargex[i]-pottarg[i]),2);
@@ -120,7 +120,7 @@ void comp_err_helm(int n, int pg, int pgt, CPX *pot, CPX *potex, CPX *pottarg,
 
   if( pgt == 2)
   {
-    for(int i=0; i<3*n; i++)
+    for(int64_t i=0; i<3*n; i++)
     {
       r = r + pow(cabs(gradtargex[i]),2);
       e = e +  pow(cabs(gradtargex[i]-gradtarg[i]),2);
